Removed http2json and folded implementation into rest_response. Fixed tests
