+++ tests/test_basics.cpp
@@ -86,17 +86,24 @@ namespace siddiqsoft
     {
         auto r1 = "https://www.siddiqsoft.com:65535/"_GET;
         EXPECT_EQ(HttpMethodType::GET, r1.getMethod());
+#if defined(DEBUG)
         EXPECT_EQ(65535, r1.uri.authority.port);
+#endif
 
         auto r2 = "https://localhost:65535/"_GET;
         EXPECT_EQ(HttpMethodType::GET, r2.getMethod());
+#if defined(DEBUG)
         EXPECT_EQ(65535, r2.uri.authority.port);
+#endif
 
         auto r3 = "https://user.name@reqbin.com:9090/echo/post/json?source=Validate::test1&param=r3"_OPTIONS;
         EXPECT_EQ(HttpMethodType::OPTIONS, r3.getMethod());
-        EXPECT_EQ(9090, r3.uri.authority.port);
 
-        nlohmann::json doc{r3};
+#if defined(DEBUG)
+        EXPECT_EQ(9090, r3.uri.authority.port);
+#endif
+
+        nlohmann::json doc {r3};
         std::cerr << "Serialized (encoded): " << r3 << std::endl;
         std::cerr << "Serialized (json'd) : " << doc.dump(2) << std::endl;
     }
